move things around to fix circularity
